Promote available balance when no booked balance
