+++ src/components/Account.vue
@@ -2,13 +2,15 @@
   <div>
     <div class="row">
       <div class="account">
-        <p class="product-name">{{ product ? product.ProductName : '' }}</p>
+        <p class="product-name">{{ product ? product.ProductName : 'Current Account' }}</p>
         <p class="sort-code-account-number">{{sortCodeAndAccountNumber}}</p>
       </div>
       <div class="balances">
-        <balance-booked v-bind:balances="balances"></balance-booked>
+        <balance-booked
+          v-bind:balances="balances"
+          v-bind:availableBalance="availableBalance"></balance-booked>
         <br />
-        <balance-available v-bind:balances="balances"></balance-available>
+        <balance-available v-bind:availableBalance="availableBalance"></balance-available>
       </div>
       <!--
       <div class="meta">
@@ -26,11 +28,58 @@
 import BalanceAvailable from './BalanceAvailable';
 import BalanceBooked from './BalanceBooked';
 
+const groupby = require('lodash.groupby');
+const sortby = require('lodash.sortby');
+
 export default {
   name: 'account',
   components: { BalanceAvailable, BalanceBooked },
   props: ['account', 'aspsp'],
+  methods: {
+    isAvailableBalance(balance) {
+      switch (balance.Type) {
+        case 'ClosingAvailable':
+        case 'InterimAvailable':
+        case 'OpeningAvailable':
+        case 'ForwardAvailable':
+          return true;
+        default:
+          return false;
+      }
+    },
+    sortDatetimeDescending(balances) {
+      return sortby(balances, b => b.DateTime).reverse();
+    },
+    bestMatchWithoutDateTime(list) {
+      const byType = groupby(list, b => b.Type);
+      if (byType.ClosingAvailable) return byType.ClosingAvailable[0];
+      if (byType.InterimAvailable) return byType.InterimAvailable[0];
+      if (byType.OpeningAvailable) return byType.OpeningAvailable[0];
+      if (byType.ForwardAvailable) return byType.ForwardAvailable[0];
+      return null;
+    },
+    bestMatch(available) {
+      const orderedByDate = this.sortDatetimeDescending(available);
+      const latestDateTime = orderedByDate[0].DateTime;
+      const byDatetime = groupby(available, b => b.DateTime);
+      const recent = byDatetime[latestDateTime];
+      if (recent.length > 1) {
+        return this.bestMatchWithoutDateTime(recent);
+      }
+      return recent[0];
+    },
+  },
   computed: {
+    availableBalances() {
+      return this.balances.filter(this.isAvailableBalance);
+    },
+    availableBalance() {
+      const available = this.availableBalances;
+      if (available.length > 0) {
+        return this.bestMatch(available);
+      }
+      return null;
+    },
     sortCodeAndAccountNumber() {
       if (this.account.Account.SchemeName === 'SortCodeAccountNumber') {
         const sortCode1 = this.account.Account.Identification.substring(0, 2);
